// api/mssql/types/SIDET.type.ts
export interface SIDET {
    ID: number;
    JOURMODE?: string | null;
    JOURTYPE?: string | null;
    JOURDATE?: Date | null;
    BILLTYPE?: string | null;
    BILLDATE?: Date | null;
    BILLNO?: string | null;
    LINE?: number | null;
    ITEMNO?: string | null;
    BCODE?: string | null;
    PCODE?: string | null;
    MCODE?: string | null;
    DETAIL?: string | null;
    WHNUMBER?: string | null;
    LOCATION1?: string | null;
    STATUS?: number | null;
    SERIAL?: string | null;
    TAXIC?: string | null;
    EXMPT?: string | null;
    ISVAT?: string | null;
    QTY?: number | null;
    UI?: string | null;
    MTP?: number | null;
    PRICE?: number | null;
    XPRICE?: number | null;
    DISCNT1?: number | null;
    DISCNT2?: number | null;
    DISCNT3?: number | null;
    DISCNT4?: number | null;
    DED?: string | null;
    VAT?: number | null;
    AMOUNT?: number | null;
    CHGAMT?: string | null;
    ACCTNO?: string | null;
    PAID?: string | null;
    ACCT_NO?: string | null;
    DONE?: string | null;
    CANCELED?: string | null;
  }